Issue #216 - Integration with apache.commons.fileupload enhancement - Condensing the classes into a single class
